Add files via upload
  Successfully utilized Natural Language Processing (NLP) techniques to categorize a dataset comprising over 5000 SMS messages into two distinct groups: "spam" and "ham." Through this process, our model has effectively determined whether each message falls into the category of unwanted or irrelevant ("spam") or relevant and desired ("ham").
 Achievement is highlighted by an accuracy rate of 97%, which underscores the accuracy of our model's predictions in relation to the entire dataset. This accomplishment demonstrates the effectiveness of our approach in accurately distinguishing between these two classes of SMS messages.

In addition to accuracy, I have also employed a comprehensive evaluation approach by generating a classification report. This report includes metrics such as precision, recall, and the F1-score. These metrics provide a more nuanced understanding of our model's performance, particularly when confronted with situations involving imbalanced datasets or varying types of errors.

Overall, our accomplishment of achieving a 97% accuracy in categorizing SMS messages is noteworthy. However, we continue to analyze the insights provided by the classification report to refine our model even further. This allows us to ensure that our approach aligns with the specific objectives and requirements of our application, potentially leading to further enhancements and refinements in the future.